+++ src/mock_io.c
@@ -43,9 +43,9 @@ volatile uint8_t virtualTCNT2;
 
 // uart0
 volatile uint16_t virtualUBRR0;
-volatile uint16_t virtualUCSR0A;
-volatile uint16_t virtualUCSR0B;
-volatile uint16_t virtualUDR0;
+volatile uint8_t virtualUCSR0A;
+volatile uint8_t virtualUCSR0B;
+volatile uint8_t virtualUDR0;
 
 uint8_t parity_even_bit(const uint8_t v) {
     uint8_t ret = v;
